--- polaris-plugins/polaris-plugins-configfilter/configfilter-crypto/src/main/java/com/tencent/polaris/plugins/configfilter/crypto/config/CryptoConfig.java
@@ -1,47 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making Polaris available.
- *
- * Copyright (C) 2019 THL A29 Limited, a Tencent company. All rights reserved.
- *
- * Licensed under the BSD 3-Clause License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package com.tencent.polaris.plugins.configfilter.crypto.config;
-
-import com.tencent.polaris.api.config.verify.Verifier;
-
-/**
- * @author fabian4
- * @date 2023/1/13
- */
-public class CryptoConfig implements Verifier {
-
-    private String type;
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    @Override
-    public void verify() {
-
-    }
-
-    @Override
-    public void setDefault(Object defaultObject) {
-
-    }
-}
