+++ polaris-plugins/polaris-plugins-configfilter/configfilter-crypto/src/main/resources/META-INF/services/com.tencent.polaris.api.plugin.crypto.ConfigFilterCrypto
@@ -1,0 +1,1 @@
+com.tencent.polaris.plugins.configfilter.crypto.AESConfigFilterCrypto
